% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_to_letter.R
\name{colnum_to_letter}
\alias{colnum_to_letter}
\title{Column to letter}
\usage{
colnum_to_letter(column_number)
}
\arguments{
\item{column_number}{Given a number, this function will tell you the corresponding alphabatized column}
}
\value{
A sequence of letters corresponding to the column number e.g., 1 = A, 26 = Z, 27 = AA
}
\description{
Column to letter
}
