% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_artist_name.R
\name{get_artist_name}
\alias{get_artist_name}
\title{Get Monthly Listeners}
\usage{
get_artist_name(artist_code)
}
\arguments{
\item{artist_code}{unique 22 alpha-numeric code to identify spotify artists}
}
\value{
A character of the artist's name
}
\description{
Get Monthly Listeners
}
