% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_monthly_listeners}
\alias{get_monthly_listeners}
\title{Get Monthly Listeners}
\usage{
get_monthly_listeners(artist_code)
}
\arguments{
\item{artist_code}{unique 22 alpha-numeric code to identify spotify artists}
}
\value{
An integer represetning total monthly streams of an artists
}
\description{
Get Monthly Listeners
}
