% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{league_baseline}
\alias{league_baseline}
\title{league Baseline}
\format{
A list containing eight dataframs each with: \code{artist} and
\code{artist_code}.
}
\usage{
league_baseline
}
\description{
The original baseline data for league standings
}
\keyword{datasets}
