% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_monthly_streams.R
\name{list_get_monthly_listerns}
\alias{list_get_monthly_listerns}
\title{list get montly listeners}
\usage{
list_get_monthly_listerns(dataframe, artist_code_column = 2)
}
\arguments{
\item{dataframe}{A dataframe that contains artist stream}

\item{artist_code_column}{the column in which the artist codes are stored}
}
\value{
An vector represetning total monthly streams of an artists
}
\description{
list get montly listeners
}
