% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_spotify_link}
\alias{parse_spotify_link}
\title{Parse Spotify Links}
\usage{
parse_spotify_link(link)
}
\arguments{
\item{link}{}
}
\value{
A character of an artist's spotify code
}
\description{
Parse Spotify Links
}
