% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_streams.R
\name{update_streams}
\alias{update_streams}
\title{Update streams}
\usage{
update_streams(list)
}
\arguments{
\item{list}{A list containing multiple dataframes that house artist codes}
}
\value{
A list with a new column indciating todays mohtly stream count
}
\description{
Update streams
}
